Create skeleton query extension, refactor CodeMirror related files
